Simplify navigation and header by introducing loader functions
