+++ src/js/App/Header/HeaderTests/Header.test.js
@@ -1,13 +1,15 @@
 import React from 'react';
 import toJson from 'enzyme-to-json';
-import { mount } from 'enzyme';
+import { shallow, mount } from 'enzyme';
 import configureStore from 'redux-mock-store';
 import { Provider } from 'react-redux';
-import AppSwitcher from './AppSwitcher';
+import Header from '../Header';
+import UnauthedHeader from '../UnAuthtedHeader';
 
-describe('AppSwitcher', () => {
+describe('Header', () => {
     let initialState;
     let mockStore;
+
     beforeEach(() =>{
         mockStore = configureStore();
         initialState = ({ chrome: {
@@ -15,11 +17,18 @@ describe('AppSwitcher', () => {
             activeLocation: 'someLocation'
         } });
     });
-    it('should render correctly', () =>{
+    it('should render correctly', () => {
         const store = mockStore(initialState);
         const wrapper = mount(<Provider store={store}>
-            <AppSwitcher currentApp='Red Hat Insights'/>
+            <Header currentApp='Red Hat Insights'/>
         </Provider>);
         expect(toJson(wrapper, { mode: 'deep' })).toMatchSnapshot();
+    });
+});
+
+describe('unauthed', () => {
+    it('should render correctly', () => {
+        const wrapper = shallow(<UnauthedHeader/>);
+        expect(toJson(wrapper)).toMatchSnapshot();
     });
 });
