+++ src/js/App/Header/Loader.js
@@ -1,0 +1,5 @@
+import React from 'react';
+
+const HeaderLoader = () => <React.Fragment />;
+
+export default HeaderLoader;
