+++ src/js/App/Header/index.js
@@ -1,13 +1,25 @@
-import React, { Fragment } from 'react';
-import Brand from './Brand';
-import Tools from './Tools';
+import React, { lazy, Suspense } from 'react';
+import { Provider } from 'react-redux';
+import { render } from 'react-dom';
+import HeaderLoader from './Loader';
+import { spinUpStore } from '../../redux-config';
 
-const Header = () => {
-    return (
-        <Fragment>
-            <Brand />
-            <Tools />
-        </Fragment>
-    );};
+const Header = lazy(() => import(/* webpackChunkName: "Sidenav" */ './Header'));
 
-export default Header;
+export const headerLoader = () => {
+    const { store } = spinUpStore();
+    if (document.querySelector('header')) {
+        render(
+            <Provider store={store}>
+                <Suspense fallback={<HeaderLoader />}>
+                    <Header />
+                </Suspense>
+            </Provider>,
+            document.querySelector('header')
+        );
+    }
+
+    if (window.insights.chrome.isPenTest()) {
+        document.querySelector('header').classList.add('ins-c-pen-test');
+    }
+};
