+++ src/js/App/Sidenav/AppSwitcher.js
@@ -6,12 +6,10 @@ import { Dropdown } from '@patternfly/react-core/dist/js/components/Dropdown/Dro
 import { DropdownItem } from '@patternfly/react-core/dist/js/components/Dropdown/DropdownItem';
 import { DropdownToggle } from '@patternfly/react-core/dist/js/components/Dropdown/DropdownToggle';
 import { CaretDownIcon } from '@patternfly/react-icons/dist/js/icons/caret-down-icon';
-import { Skeleton, SkeletonSize } from '@redhat-cloud-services/frontend-components/components/cjs/Skeleton';
-import { connect } from 'react-redux';
 
 import './AppSwitcher.scss';
 
-const AppSwitcher = ({ currentApp, globalNav }) => {
+const AppSwitcher = ({ currentApp }) => {
 
     const [isOpen, setIsOpen] = useState(false);
 
@@ -41,32 +39,25 @@ const AppSwitcher = ({ currentApp, globalNav }) => {
     };
 
     return (
-        <section className={`ins-c-app-switcher${globalNav ? '' : '--loading'}`}>
-            {
-                globalNav ? (
-                    <Dropdown
-                        isPlain
-                        className='ins-c-app-switcher__dropdown'
-                        onSelect={() => setIsOpen(!isOpen)}
-                        toggle={
-                            <DropdownToggle id="toggle-id" onToggle={() => setIsOpen(!isOpen)} toggleIndicator={CaretDownIcon}>
-                                { currentApp }
-                            </DropdownToggle>
-                        }
-                        isOpen={isOpen}
-                        dropdownItems={dropdownItems(currentApp)}
-                    />
-                ) : <Skeleton size={SkeletonSize.lg} className="ins-m-dark"/>
-            }
+        <section className="ins-c-app-switcher">
+            <Dropdown
+                isPlain
+                className='ins-c-app-switcher__dropdown'
+                onSelect={() => setIsOpen(!isOpen)}
+                toggle={
+                    <DropdownToggle id="toggle-id" onToggle={() => setIsOpen(!isOpen)} toggleIndicator={CaretDownIcon}>
+                        { currentApp }
+                    </DropdownToggle>
+                }
+                isOpen={isOpen}
+                dropdownItems={dropdownItems(currentApp)}
+            />
         </section>
     );
 };
 
-export default connect(({ chrome: { globalNav } }) => ({ globalNav }))(AppSwitcher);
+export default AppSwitcher;
 
 AppSwitcher.propTypes = {
-    currentApp: PropTypes.string,
-    globalNav: PropTypes.shape({
-        [PropTypes.string]: PropTypes.any
-    })
+    currentApp: PropTypes.string
 };
