+++ src/js/App/Sidenav/Navigation.js
@@ -7,7 +7,6 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { appNavClick, clearActive } from '../../redux/actions';
 import NavigationItem from './NavigationItem';
-import { Skeleton, SkeletonSize } from '@redhat-cloud-services/frontend-components/components/cjs/Skeleton';
 
 import './Navigation.scss';
 
@@ -35,6 +34,7 @@ const insightsLinks = {
     }
 };
 
+// TODO: refactor this funcion
 export class Navigation extends Component {
     constructor(props) {
         super(props);
@@ -94,7 +94,7 @@ export class Navigation extends Component {
             <Nav onSelect={this.onSelect} aria-label="Insights Global Navigation" data-ouia-safe="true">
                 <NavList>
                     {
-                        settings ? settings.map((item, key) => {
+                        settings?.map((item, key) => {
                             if (!(item.disabled_on_stable && window.location.pathname.indexOf('/beta') === -1)) {
                                 if (item.subItems) {
                                     return <NavExpandable
@@ -136,14 +136,7 @@ export class Navigation extends Component {
                                     />;
                                 }
                             }
-                        }) : [...new Array(4)].map((_i, key) => (
-                            <NavigationItem
-                                key={key}
-                                title={<a className="ins-c-skeleton__link">
-                                    <Skeleton size={SkeletonSize.lg} className="ins-m-dark"/>
-                                </a>}
-                            />
-                        ))
+                        })
                     }
                     { activeLocation === 'insights' &&
                         Object.entries(insightsLinks).map(
